refactor: memo, logic 분리
