+++ src/components/LeftRight/index.ts
@@ -1,7 +1,6 @@
 export { default as Arrows } from "./Arrows";
 export { default as Combo } from "./Combo";
 export { default as HorizontalLine } from "./HorizontalLine";
-export { default as LastArrowHighlighter } from "./LastArrowHighlighter";
 export { default as MaxCombo } from "./MaxCombo";
 export { default as MobilePad } from "./MobilePad";
 export { default as Score } from "./Score";
