+++ src/pages/game/left-right.tsx
@@ -1,9 +1,6 @@
 import { Center, Flex } from "@chakra-ui/react";
 import { motion } from "framer-motion";
-import { useAtomValue, useSetAtom } from "jotai";
-import { useEffect } from "react";
 
-import { correctAtom, lastArrowAtom, shakeAtom, wrongAtom } from "../../atoms/left-right";
 import HangHeart from "../../components/common/HangHeart";
 import {
   Arrows,
@@ -16,31 +13,10 @@ import {
   VerticalLine,
 } from "../../components/LeftRight";
 import { LAYER } from "../../constants";
+import useLeftRightGame from "../../hooks/useLeftRightGame";
 
 const LeftRightGame = () => {
-  const lastArrow = useAtomValue(lastArrowAtom);
-  const shake = useAtomValue(shakeAtom);
-
-  const correct = useSetAtom(correctAtom);
-  const wrong = useSetAtom(wrongAtom);
-
-  useEffect(() => {
-    const keyDownHandler = (e: KeyboardEvent) => {
-      if (e.key !== "ArrowLeft" && e.key !== "ArrowRight") return;
-
-      if (
-        (e.key === "ArrowLeft" && lastArrow.direction === "left") ||
-        (e.key === "ArrowRight" && lastArrow.direction === "right")
-      ) {
-        correct();
-      } else {
-        wrong();
-      }
-    };
-
-    window.addEventListener("keydown", keyDownHandler);
-    return () => window.removeEventListener("keydown", keyDownHandler);
-  }, [correct, lastArrow.direction, wrong]);
+  useLeftRightGame();
 
   return (
     <Center position="relative" backgroundColor="primary.900">
@@ -67,10 +43,6 @@ const LeftRightGame = () => {
             width: "75%",
             height: "50%",
           }}
-          animate={{
-            translateX: shake ? [-5, 5, -5, 5, 0] : 0,
-          }}
-          transition={{ type: "spring", stiffness: 500, damping: 30 }}
         >
           <Arrows />
           <Score />
